% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_login.R
\name{aws_configure}
\alias{aws_configure}
\title{Function to create SSO configuration}
\usage{
aws_configure(env)
}
\arguments{
\item{env}{environment variables}
}
\description{
Create ~/.aws/config with requirements to SSO to DataBrew AWS Accounts
}
