% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_get.R
\name{aws_get_data_catalog}
\alias{aws_get_data_catalog}
\title{get s3 data catalog for databrew files}
\usage{
aws_get_data_catalog(bucket, namespace_bucket = TRUE, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{...}{additional parameter from S3 get object}
}
\value{
returns databrew.org data catalog
}
\description{
get s3 data catalog for databrew files
}
