% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_get.R
\name{aws_get_object}
\alias{aws_get_object}
\title{get object in s3}
\usage{
aws_get_object(bucket, key, namespace_bucket = TRUE, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{key}{s3 object key}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{...}{additional parameter from S3 get object}
}
\value{
list of object s3 metadata and file location
}
\description{
get object in s3
}
