% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_table.R
\name{aws_get_table}
\alias{aws_get_table}
\title{Wrapper function to retrieving object into table format data}
\usage{
aws_get_table(bucket, key, namespace_bucket = TRUE, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{key}{s3 object key}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{...}{additional parameter from S3 get object}
}
\value{
tibble data-frame
}
\description{
Wrapper function to retrieving object into table format data
}
