% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_table.R
\name{aws_get_table_ts}
\alias{aws_get_table_ts}
\title{Wrapper function to retrieving object into time-series dataa}
\usage{
aws_get_table_ts(bucket, key, namespace_bucket = TRUE, date_range = NULL, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{key}{object key}

\item{namespace_bucket}{boolean for bucket namespacing, set to FALSE to override namespace}

\item{date_range:}{date range min/max of date partition}
}
\value{
tibble dataframe with run-date index
}
\description{
Wrapper function to retrieving object into time-series dataa
}
