% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_login.R
\name{aws_login}
\alias{aws_login}
\title{Utility Function to login to DataBrew AWS.}
\usage{
aws_login(
  role_name = "cloudbrewr-aws-role",
  profile_name = "dbrew-cloudbrewr-role",
  access_key = NULL,
  secret_access_key = NULL,
  session_token = NULL,
  account_region = "us-east-1",
  pipeline_stage = "production"
)
}
\arguments{
\item{role_name}{set role name based on SSO portal, this will be different based on different users but default values will be set to cloudbrewr-aws-role}

\item{profile_name}{set profile name to save your SSO configuration locally, this will be different based on different users}

\item{access_key}{retrieve from portal for access key access}

\item{secret_access_key}{retrieve from portal for secret access key}

\item{session_token}{retrieve fro portal}

\item{account_region}{default to us-east-1}

\item{pipeline_stage}{(DataBrew internal use for CI/CD) choose production/develop stage}
}
\value{
metadata of AWS STS authentication (Account, Role)
}
\description{
If user is in an interactive session, generate all the required credentials to run SSO.
If user is running in Terminal / bash / workflow / VM prompt to export temporary credentials
}
