% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aws_namespace}
\alias{aws_namespace}
\title{Function to namespace bucket}
\usage{
aws_namespace(bucket)
}
\arguments{
\item{bucket}{databrew s3 bucket}
}
\value{
namspaced buckeet
}
\description{
This function is used to namespace bucket based on pipelines stages
}
