% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_store.R
\name{aws_s3_bulk_store}
\alias{aws_s3_bulk_store}
\title{Bulk Store Object in DataBrew S3}
\usage{
aws_s3_bulk_store(
  bucket,
  prefix = NULL,
  namespace_bucket = TRUE,
  target_dir = NULL,
  ...
)
}
\arguments{
\item{bucket}{s3 bucket to store data}

\item{prefix}{(optional) folder prefix to store data}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{target_dir}{folder to sync with aws}
}
\value{
list of object s3 metadata and file location
}
\description{
Bulk Store Object in DataBrew S3
}
