% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_create_bucket.R
\name{aws_s3_create_bucket}
\alias{aws_s3_create_bucket}
\title{Function to create S3 bucket}
\usage{
aws_s3_create_bucket(bucket, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{...}{additional parameter passed to s3 create_bucket endpoint}
}
\description{
create S3 bucket via API
}
