% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_get.R
\name{aws_s3_get_catalog}
\alias{aws_s3_get_catalog}
\title{Get DataBrew S3 object catalog}
\usage{
aws_s3_get_catalog(bucket, namespace_bucket = TRUE, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{...}{additional parameter from S3 get object}
}
\value{
tibble of S3 file catalog
}
\description{
Get catalog of files used in S3
}
