% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_get.R
\name{aws_s3_get_header}
\alias{aws_s3_get_header}
\title{Get Metadata Headers}
\usage{
aws_s3_get_header(bucket, key, namespace_bucket = TRUE)
}
\arguments{
\item{bucket}{s3 bucket}

\item{key}{s3 object key}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{...}{additional parameter from S3 get object}
}
\value{
list of object s3 metadata and file location
}
\description{
Get Metadata Headers
}
