% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_table.R
\name{aws_s3_get_table_ts}
\alias{aws_s3_get_table_ts}
\title{Function to retrieve S3 folder-partitioned table into time-series data}
\usage{
aws_s3_get_table_ts(
  bucket,
  key,
  namespace_bucket = TRUE,
  date_range = NULL,
  ...
)
}
\arguments{
\item{bucket}{s3 bucket}

\item{key}{object key}

\item{namespace_bucket}{boolean for bucket namespacing, set to FALSE to override namespace}

\item{date_range}{date range min/max of date partition}

\item{...}{argument on s3 api on list_objects_v2}
}
\value{
tibble dataframe with run-date index
}
\description{
Wrapper to bind folder-partitioned table into time-series, indexed by run_date
}
