% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_login.R
\name{aws_sso_authenticate}
\alias{aws_sso_authenticate}
\title{Function to initiate SSO authentication}
\usage{
aws_sso_authenticate(profile_name, account_id)
}
\arguments{
\item{profile_name}{the profile name set to prod/dev}
}
\description{
Instantiate SSO connection using internet browser
}
