% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_login.R
\name{aws_sso_login}
\alias{aws_sso_login}
\title{Function to login to aws
use this on interactive session
to work on local testing
NOTE: USE THIS FOR LOCAL TESTING ONLY}
\usage{
aws_sso_login(prod = TRUE)
}
\arguments{
\item{prod}{prod/dev choose AWS account}
}
\description{
Function to login to aws
use this on interactive session
to work on local testing
NOTE: USE THIS FOR LOCAL TESTING ONLY
}
