% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_cloudbrewr_stage_env_variables}
\alias{call_cloudbrewr_stage_env_variables}
\title{Function to get prod/dev stage environment variables,}
\usage{
call_cloudbrewr_stage_env_variables(pipeline_stage)
}
\arguments{
\item{pipeline_stage}{choose production/develop stage}
}
\description{
Retrieve environment variables used for
S3 bucket namespace and profile name to
your environment variables.
}
