% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cloudbrewr_cache}
\alias{check_cloudbrewr_cache}
\title{Function to check file cache}
\usage{
check_cloudbrewr_cache(etag, cache = "~/.cloudbrewr_cache")
}
\arguments{
\item{etag}{S3 etag - this is unique identifier of a file}
}
\value{
boolean
}
\description{
This function is used to check downloaded files in cache
}
