% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_cache}
\alias{pull_cache}
\title{Function to pull from cache}
\usage{
pull_cache(cache = "~/.cloudbrewr_cache")
}
\arguments{
\item{cache}{desired cache location, set to cloudbrewr_cache}
}
\value{
cachem object
}
\description{
This function is used to instantiate cache filesystem
}
