% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_login.R
\name{aws_configure}
\alias{aws_configure}
\title{Function to create SSO configuration}
\usage{
aws_configure(
  profile_name,
  role_name,
  account_region = "us-east-1",
  account_id = "354598940118"
)
}
\arguments{
\item{env}{environment variables}
}
\description{
Create ~/.aws/config with requirements to SSO to DataBrew AWS Accounts
}
