% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_get.R
\name{aws_s3_bulk_get}
\alias{aws_s3_bulk_get}
\title{Bulk Get Object in DataBrew S3}
\usage{
aws_s3_bulk_get(
  bucket,
  prefix = NULL,
  namespace_bucket = TRUE,
  output_dir = "~/.cloudbrewr_cache",
  ...
)
}
\arguments{
\item{bucket}{s3 bucket}

\item{prefix}{folder prefix}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{output_dir}{output directory from aws_s3_get parameter, this will be the destination if write_cache is set to \code{TRUE}}

\item{...}{additional parameter from S3 get object}
}
\value{
list of object s3 metadata and file location
}
\description{
Bulk Get Object in DataBrew S3
}
