% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_get.R
\name{aws_s3_get_object_version_history}
\alias{aws_s3_get_object_version_history}
\title{Get object version history in s3}
\usage{
aws_s3_get_object_version_history(bucket, key, namespace_bucket = TRUE, ...)
}
\arguments{
\item{bucket}{s3 bucket}

\item{key}{s3 object key}

\item{namespace_bucket}{boolean to create namespace bucket, set to FALSE to override bucket namespace}

\item{...}{additional parameter from S3 get object}
}
\value{
tibble of S3 object version history
}
\description{
Get object version change history
}
