% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_cloudbrewr_stage_env_variables}
\alias{call_cloudbrewr_stage_env_variables}
\title{Function to get prod/dev stage environment variables,
use this to retrieve environment variables used for
S3 bucket namespace and profile name to your environment variables.}
\usage{
call_cloudbrewr_stage_env_variables(stage)
}
\arguments{
\item{stage}{values \code{prod}, \code{dev}}
}
\description{
Function to get prod/dev stage environment variables,
use this to retrieve environment variables used for
S3 bucket namespace and profile name to your environment variables.
}
